Move Bluez
