+++ src/features/ble/bluez/generate.sh
