+++ src/features/ble/bluez/org.bluez-gatt-class.ts
